Add files via upload
Percentage body fat (underwater) is often looked at by health professionals as an
indicator of an individual’s health. Although, it is a useful data point, calculating (underwater)
percentage body fat is expensive and often inconvenient. As such, it would be useful to create a
model to predict (underwater) percentage body fat